move Icon component to Icon.tsx
